+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # MMM-Fuel Changelog
 
+## [2.3.0]
+
+MagicMirror² version >= 2.15.0 required.
+
+### Added
+
+* Provider: `autoblog`
+* Config option: `showBrand`
+* Integrated MagicMirror logger on server side
+* Dependency: `node-html-parser`
+* Config option `excludeStationIds` to exclude gas stations from radius (Tankerkönig only)
+
+### Changed
+
+* Updated Github config files
+* Updated dependencies
+
+### Removed
+
+* Dependency: `fs-extra`
+* Dependency: `geolib`
+
 ## [2.2.1]
 
 ### Fixed